Fix AdvCiv AI issue with vote calculation for diplo victory
Vassal votes were being disregarded. This had been working correctly before, I probably broke it in v0.97. [advc.115b]

+++ CvGameCoreDLL/WarUtilityAspect.h
@@ -212,7 +212,7 @@ private:
 	double progressRatingDiplomacy();
 	void addConquestsByPartner(std::map<int,double>& r, AttitudeTypes attThresh,
 			double weight);
-	double votesToGo;
+	int votesToGo;
 	bool enoughVotes;
 };
